Changed for new version (v_1.10) of lnfl.
For H2O the lower state energies with value of  -1. have been replaced
with a value of -300. cm-1 outside of lnfl in the linefile (the negative
value of the lower state energy indicates that the value is not known).
This value in now changed in the line file to achieve a more probable
temperature correction.  LNFL has been changed so that it DOES NOT
change this E" value and passes the -300 directly to LBLRTM.